requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.9.24
charset-normalizer==2.1.1
cloudinary==1.30.0
Django==4.1.3
django-cloudinary-storage==0.3.0
django-environ==0.9.0
django-filter==22.1
djangorestframework==3.14.0
idna==3.4
jsonfield==3.1.0
mysqlclient==2.1.1
Pillow==9.3.0
PyMySQL==1.0.2
pytz==2022.6
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
tzdata==2022.6
urllib3==1.26.13
whitenoise==6.2.0
